Update with changes in HEAD, to revision 10520
git-svn-id: file:///home/xapian-svn/xapian/branches/matchspy@10521 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/net/remoteconnection.cc
@@ -554,7 +554,7 @@ RemoteConnection::get_message_chunk(string &result, size_t at_least,
 
     read_at_least(at_least, end_time);
 
-    size_t retlen(size_t(min(off_t(buffer.size()), chunked_data_left)));
+    size_t retlen(min(off_t(buffer.size()), chunked_data_left));
     result.append(buffer, 0, retlen);
     buffer.erase(0, retlen);
     chunked_data_left -= retlen;
